refactor(all): move binding-mode and binding-flags to binding
