+++ src/runtime/aurelia.ts
@@ -1,8 +1,8 @@
-import { BindingFlags } from './binding/binding-flags';
 import { ICustomElement } from './templating/custom-element';
 import { PLATFORM } from '../kernel/platform';
 import { IContainer, DI, IRegistry } from '../kernel/di';
 import { IRenderingEngine } from './templating/rendering-engine';
+import { BindingFlags } from './binding/binding';
 
 export interface ISinglePageApp {
   host: any,
