--- src/runtime/binding/binding-mode.ts
@@ -1,6 +1,0 @@
-export enum BindingMode {
-  oneTime  = 0b00,
-  toView   = 0b01,
-  fromView = 0b10,
-  twoWay   = 0b11
-};
