+++ src/runtime/templating/resources/repeater.ts
@@ -10,14 +10,12 @@ import { ITaskQueue } from '../../task-queue';
 import { IRenderSlot } from '../../templating/render-slot';
 import { IViewOwner } from '../../templating/view';
 import { IVisualFactory, IVisual } from '../../templating/visual';
-import { IScope, IOverrideContext, sourceContext } from '../../binding/binding-context';
+import { IScope, IOverrideContext } from '../../binding/binding-context';
 import { ForOfStatement } from '../../binding/ast';
-import { Binding } from '../../binding/binding';
-import { BindingMode } from '../../binding/binding-mode';
+import { Binding, BindingMode, BindingFlags } from '../../binding/binding';
 import { Immutable } from '../../../kernel/interfaces';
 import { IResourceKind, IResourceType } from '../../resource';
-import { INode, DOM } from '../../dom';
-import { BindingFlags } from '../../binding/binding-flags';
+import { INode } from '../../dom';
 import { ICustomAttribute } from '../custom-attribute';
 import { getMapObserver } from '../../binding/observation/map-observer';
 import { getSetObserver } from '../../binding/observation/set-observer';
