+++ src/runtime/templating/resources/replaceable.ts
@@ -3,7 +3,7 @@ import { IRenderSlot } from '../render-slot';
 import { IScope } from '../../binding/binding-context';
 import { inject } from '../../../kernel/di';
 import { IVisualFactory, IVisual } from '../visual';
-import { BindingFlags } from '../../binding/binding-flags';
+import { BindingFlags } from '../../binding/binding';
 
 @templateController('replaceable')
 @inject(IVisualFactory, IRenderSlot)
